tests: validate media searching
Its plainly obvious so far that recursing two directories and finding 164
items in 0m0.240s is not bad going. The bottlekneck will come from obtaining
their id3 tag data, and adding them to the database.

Currently no timing tests are in place, and won't be until some parts
of the API have been fixed up to be cancellable and report where, when,
and how many files we're encountering.