Added widget class for track listings.
This should resolve evolve-os/budgie-media-player#16 from upstream.

The BudgieTrackList class should be reusable for things like playlists, should
such a feature be implemented. It's not optimised - I think there are parts
where we destroy or reload stuff unnecessarily.

Upon further research, it seems that use of ListView is to blame. I'll work on replacing it with a TreeView/Model themed to fit.