Replaced the GtkListBox with GtkTreeView.
This should be more flexible and fast, once we reduce the rebuilding of stores/databases. I think reimplementing BudgieDB as a ListModel-type thing is a good idea at some point in the future.

The TreeView has been styled to be mostly equivalent to the old style. It's not perfect, and it uses a stock icon which might look a bit ugly depending on the theme, but it's definitely functional. I haven't timed it, but it feels a bit better to me. It's not totally smooth yet, but it's getting there.