Migrated from GDBM to SQLite3
This change should make the database easier to work with. It seems much faster, although I haven't actually profiled it. It should make searching faster and more flexible, make versioning and extending the database easier, and improve flexibility of the structure.

I also added the track number tag to the MediaInfo structure. While not strictly necessary, it was odd that it was missing. It defaults to 0, per taglib.

I realise this is a pretty big change, but it seems to be working well. The CONFIG_NAME macro has been updated to "budgie-2.db" to reflect the magnitude of this change.